RatingCompleter.java and RatingCMD.java merged to RatingCmd.java, events moved from Rating.java to PlayerListener.java
