+++ src/main/java/me/h1r0sh1m4/rating/Rating.java
@@ -1,29 +1,34 @@
 package me.h1r0sh1m4.rating;
 
-import me.h1r0sh1m4.rating.commands.ratingCMDpkg.RatingCMD;
-import me.h1r0sh1m4.rating.commands.ratingCMDpkg.RatingCompleter;
+import me.h1r0sh1m4.rating.commands.RatingCmd;
 import me.h1r0sh1m4.rating.db.MySQLDatabase;
+import me.h1r0sh1m4.rating.events.PlayerListener;
 import org.bukkit.Bukkit;
 import org.bukkit.configuration.file.FileConfiguration;
-import org.bukkit.event.EventHandler;
-import org.bukkit.event.Listener;
-import org.bukkit.event.player.PlayerJoinEvent;
+import org.bukkit.event.HandlerList;
 import org.bukkit.plugin.java.JavaPlugin;
 
-public final class Rating extends JavaPlugin implements Listener {
+public final class Rating extends JavaPlugin {
+
     private MySQLDatabase database;
 
     @Override
     public void onEnable() {
+
         saveDefaultConfig();
+
         database = connectDB();
-        getCommand("rating").setExecutor(new RatingCMD(database));
-        getCommand("rating").setTabCompleter(new RatingCompleter());
-        Bukkit.getPluginManager().registerEvents(this, this);
+
+        RatingCmd ratingCmd = new RatingCmd(database);
+        getCommand("rating").setExecutor(ratingCmd);
+        getCommand("rating").setTabCompleter(ratingCmd);
+
+        Bukkit.getPluginManager().registerEvents(new PlayerListener(database), this);
     }
 
     @Override
     public void onDisable() {
+        HandlerList.unregisterAll(this);
         database.close();
     }
 
@@ -35,15 +40,5 @@ public final class Rating extends JavaPlugin implements Listener {
         String username = config.getString("database.username");
         String password = config.getString("database.password");
         return new MySQLDatabase(host, port, dbName, username, password);
-    }
-
-    @EventHandler
-    private void onPlayerJoin(PlayerJoinEvent event) {
-        String playerName = event.getPlayer().getName();
-        if (!database.exists(playerName)) {
-            database.addPlayer(playerName);
-        }
-        event.getPlayer().sendMessage(database.getHistory(playerName));
-        database.addHistory(playerName, "", true);
     }
 }
